[Feat] MemberServiceImpl class -> validator 적용 및 주석 작성

+++ src/main/java/Teemo/Teemo_backend/service/MemberServiceImpl.java
@@ -37,15 +37,39 @@ public class MemberServiceImpl implements MemberService {
         Gender gender = request.getGender();
 
         /**
-         * [제약조건]
-         * 1. 빈 값 확인
-         * 2. 이메일 중복 확인
-         * 3. 닉네임 중복 확인
-         * 4. 이메일 길이 조건 확인 : [5 ~ 100] 자 사이의 길이를 가져야 한다. (5자, 100자 포함)
-         * 5. 비밀번호 길이 조건 확인 : [15] 자 길이를 가져야 한다.
-         * 6. 닉네임 길이 조건 확인 : [2 ~ 10] 자 사이의 길이를 가져야 한다. (경계값 포함)
-         * 7. 생년월일 조건 확인 : ( 금일 기준 100 년 전 ~ 금일 ) 사이의 값을 가져야 한다. (경계값 미포함)
+         * [체크리스트]
+         * 1. 이메일 중복 확인
+         * 2. 닉네임 중복 확인
+         * 3. 이메일 길이 조건 확인 : [5 ~ 100] 자 사이의 길이를 가져야 한다. (5자, 100자 포함)
+         * 4. 비밀번호 길이 조건 확인 : [64] 자 길이를 가져야 한다.
+         * 5. 닉네임 길이 조건 확인 : [2 ~ 10] 자 사이의 길이를 가져야 한다. (경계값 포함)
+         * 6. 생년월일 이 유효한지 확인:
+         *  6-1. "oooo-oo-oo" 형식으로 전달되어야 한다.
+         *  6-2. 금일 기준 100 년 전 ~ 금일 ) 사이의 값을 가져야 한다. (경계값 미포함)
          */
+        Member find = memberRepository.findByEmail(email);
+        // [체크리스트 1]
+        if(commonValidator.found(find)) // 값이 있으면 안됨,.
+            throw new CustomInvalidValueException("email", "이미 등록된 이메일 입니다.");
+        find = memberRepository.findByNickname(nickname);
+        // [체크리스트 2]
+        if(commonValidator.found(find)) // 있으면 오류
+            throw new CustomInvalidValueException("nickname","이미 등록된 닉네임 입니다.");
+        // [체크리스트 3]
+        if(!memberValidator.checkEmailLength(email))
+            throw new CustomInvalidValueException("email","이메일은 5자 에서 100자 이내로 입력되어야 합니다.");
+        // [체크리스트 4]
+        if(!memberValidator.checkPasswordLength(password))
+            throw new CustomInvalidValueException("password","비밀번호는 64자 길이를 가져야 합니다.");
+        // [체크리스트 5]
+        if(!memberValidator.checkNicknameLength(nickname))
+            throw new CustomInvalidValueException("nickname","닉네임은 2자 에서 10자 이내로 입력되어야 합니다.");
+        // [체크리스트 6-1]
+        if(!memberValidator.checkBirthdayExpression(birthday))
+            throw new CustomInvalidValueException("birthday","생년월일의 형식이 잘못 되었습니다.");
+        // [체크리스트 6-2]
+        if(!memberValidator.checkBirthdayAssertion(birthday))
+            throw new CustomInvalidValueException("birthday","생년월일의 범위가 잘못되었습니다.");
         Member member = new Member(email,password,nickname,birthday,gender);
         memberRepository.save(member);
     }
@@ -53,12 +77,13 @@ public class MemberServiceImpl implements MemberService {
     @Override
     public Member find(Long memberId) {
         /**
-         * [제약조건]
-         * 1. 빈 값인지 확인
-         * 2. 아이디로 유저를 찾지 못한 경우
+         * [체크리스트]
+         * 1. 유효한 사용자인지 확인
          */
-
         Member member = memberRepository.findById(memberId);
+        // [체크리스트 1]
+        if(!commonValidator.found(member)) // 없으면 오류 반환.
+            throw new CustomInvalidValueException("memberId","회원이 식별되지 않습니다.");
         return member;
     }
 
@@ -66,16 +91,18 @@ public class MemberServiceImpl implements MemberService {
     @Transactional
     public void update(MemberUpdateRequest request) {
         /**
-         * [제약조건]
-         * 1. 빈 값인지 확인
-         * 2. 아이디로 유저를 찾지 못한 경우
+         * [체크리스트]
+         * 1. 유효한 사용자인지 확인
          */
-        Long id = request.getId();
+        Long id = request.getMemberId();
+        Member member = memberRepository.findById(id);
+        // [체크리스트 1]
+        if(!commonValidator.found(member)) // 없으면 오류 반환.
+            throw new CustomInvalidValueException("memberId","회원이 식별되지 않습니다.");
+
         String email = request.getEmail();
         String password = request.getPassword();
         String nickname = request.getNickname();
-
-        Member member = memberRepository.findById(id);
         member.updateAccountInfo(email,password,nickname);
     }
 
@@ -83,25 +110,35 @@ public class MemberServiceImpl implements MemberService {
     @Transactional
     public void remove(Long memberId) {
         /**
-         * [제약조건]
-         * 1. 빈 값인지 확인
-         * 2. 아이디로 유저를 찾지 못한 경우
+         * [체크리스트]
+         * 1. 유효한 사용자인지 확인
          */
         Member member = memberRepository.findById(memberId);
+        // [체크리스트 1]
+        if(!commonValidator.found(member)) // 없으면 오류 반환.
+            throw new CustomInvalidValueException("memberId","회원이 식별되지 않습니다.");
         memberRepository.delete(member);
     }
 
     @Override
     public Member login(MemberLoginRequest request) {
+        /**
+         * [체크리스트]
+         * 1. 유효한 사용자인지 확인
+         * 2. 비밀번호가 일치하는지 확인
+         * 3. 기간이 만료된 회원인지 확인 (-> 완료되었다면 삭제)
+         */
         String email = request.getEmail();
         String password = request.getPassword();
         Member findMember = memberRepository.findByEmail(email);
-
+        // [체크리스트 1]
         if(!commonValidator.found(findMember))
             throw new CustomInvalidValueException("email","회원이 식별되지 않습니다.");
+        // [체크리스트 2]
         if(!memberValidator.comparePassword(password,findMember.getPassword())){
             throw new CustomInvalidValueException("password","비밀번호가 틀렸습니다.");
         }
+        // [체크리스트 3]
         if(!memberValidator.checkDeadLine(LocalDate.now(),findMember.getDeletedAt())){
             memberRepository.delete(findMember);
             throw new CustomInvalidValueException("deletedAt","삭제된 회원입니다.");
@@ -111,10 +148,18 @@ public class MemberServiceImpl implements MemberService {
 
     @Override
     public void logout(Long memberId) {
+        /**
+         * [체크리스트]
+         * 1. 유효한 사용자인지 확인
+         *  1-1. 사용자의 역할이 "HOST" 라면 Tag 를 제거해준다. (Tag 제거가 안되면, 오류 반환)
+         *  1-2. 사용자의 역할이 "GUEST" 라면 Tag 구독을 취소해준다. (Tag 구독 취소가 안되면, 오류반환)
+         *
+         */
         Member findMember = memberRepository.findById(memberId);
+        // [체크리스트 1]
         if(!commonValidator.found(findMember))
             throw new CustomInvalidValueException("memberId","회원이 식별되지 않습니다.");
-
+        // [체크리스트 1-1]
         if(findMember.getRole() == Role.HOST) {
             Long tagId = findMember.getTag().getId();
             try {
@@ -123,6 +168,7 @@ public class MemberServiceImpl implements MemberService {
                 throw new CustomInvalidValueException("tagId","Tag 를 제거해야 합니다.");
             }
         }
+        // [체크리스트 1-2]
         else if(findMember.getRole() == Role.GUEST) {
             Long tagId = findMember.getTag().getId();
             try {
